+++ griptape/config/__init__.py
@@ -13,6 +13,7 @@ from .structure_config import StructureConfig
 from .openai_structure_config import OpenAiStructureConfig
 from .amazon_bedrock_structure_config import AmazonBedrockStructureConfig
 from .anthropic_structure_config import AnthropicStructureConfig
+from .google_structure_config import GoogleStructureConfig
 
 
 __all__ = [
@@ -29,4 +30,5 @@ __all__ = [
     "OpenAiStructureConfig",
     "AmazonBedrockStructureConfig",
     "AnthropicStructureConfig",
+    "GoogleStructureConfig",
 ]